Simplify channel evaluation mechanism
Previously each channel used a fixed point of pkgs as a base set, upon
which it could add its own overlays for adding more packages. This had
some problems:
- Figuring out which attributes belong to the channel and which came
  from nixpkgs is tricky, especially with package sets, as overlays
  don't export that information easily
- Overlays are already complicated, but with this you had two sets of
  overlays: Ones for nixpkgs and ones for the channel
- Unexpected things can happen if you use nixpkgs attributes that were
  intended to be used in overlays. E.g. .extend or .pkgs and co.

This commit makes this much simpler by using nixpkgs overlays for
everything. However in order for the auto-called packages to not overlay
nixpkgs packages deeply, the auto-called packages only get written to
self.flox.outputs. And they are auto-called with both self and
self.flox.outputs in scope in order to allow using the defined packages
from others.

This also gets rid of the confusing channelConfig construct, and instead
replacing it with an auto-caller specific config, which they can manage
using flox.auto.*.

Auto-callers now also need to be defined using the `auto = ` argument of
the main function, simplifying the process a bit.