Implement needsParentChannelAccess for less duplicated channel imports
Previously for every dependency between channel X and Y, channel Y would
be evaluated, even if that import could be shared between all channels
that depend on it.

This commit introduces the needsParentChannelAccess argument for a channels
default.nix, which is false by default, but can be set to true if access
to floxInternal.parentChannel is needed, which then causes the multiple
imports. But with the default of false, the channel isn't imported
multiple times.