Make nixpkgs package set aliases also aliases in channel outputs
Before this change, pkgs.pythonPackages and pkgs.python27Packages would
be treated and merged with the channels outputs separately. This was
simpler in the code, but causes problems if nixpkgs' pythonPackages
doesn't exactly correspond to python27Packages.

While this isn't the case with pythonPackages, it's like that with
beamPackages, because beamPackages.extend doesn't exist, even though
beam.packages.erlang.extend does, which the alias should point to.

While this is kind of a nixpkgs problem as well, we can fix this on the
flox side by just making beamPackages an actual alias of the
beam.packages.erlang output. Like this, we don't need to rely on nixpkgs
aliases being actual aliases.