Move flox.getSource to meta.getSource and remove floxInternal
Previously to get a source, you'd have to use

  { flox, floxInternal }:
    flox.getSource floxInternal.importingChannelArgs.name "project" {}

which is really convoluted. This was necessary because the `getSource`
function lived in the flox channel, and its own importingChannelArgs
would be the current channel where the builder is defined, not the
importing channel of the current one.

To fix this, we just move the getSource function out of the flox
channel and into the channel setup internals instead (which
coincidentally also live in the flox channel, but under the ./channel
directory).

Now the same can be achieved with

  { meta }:
    meta.getSource "project" {}

In addition, the previous floxInternal.withVerbosity was moved to
meta.withVerbosity