Replace srcpath with sourceOverrideJson
Previously srcpath was used by the `flox build` command to inject a
mutable source into the channel build. Since it didn't specify which
packages source to override with srcpath, *all* the packages were
changed to use srcpath. This worked mostly fine because only a single
package is built in the end.

Where this became a problem however is if the package built depends on
another package of the same channel. Previously srcpath would be used
for both of them, which means the dependency would obviously fail
because it was built with the source of another package.

To fix this, srcpath is changed to sourceOverrideJson, which allows
specifying the exact project a source should be overridden for. In
addition to fixing above problem, this also allows specifying multiple
different mutable sources at once.