Prevent the super attribute from overriding the flox and co attributes
Previously, if a package was declared under attribute `flox`, `meta` or
co., it would cause its definition to receive the `flox` argument
pointing to `super.flox` (from nixpkgs, if it existed). This prevented
the definition from being able to refer to the `flox` builders.

We fix this by not allowing access to `super.flox` in case it's an
attribute that shouldn't be overridable