Change _floxFile into _floxPath
_floxPath is strictly more powerful than _floxFile, because it allows
you to easily distinguish between whether the path is a file or a
directory by just checking it. If the file is wanted and it is a
directory, a `default.nix` can just be appended.

This corresponds more directly to how floxpkgs imports paths. The
returned _floxPath contains the full extents of what the attribute is
meant to be built from.