Propagated package set versions and enforcement
This implements an internal mechanism for propagating package set versions
throughout channels and enforcing their use.

With this commit, if a package definition needs access to e.g. python
packages, it will _have_ to use the `pythonPackages` attribute, attributes
that specify a version like `python38Packages` are unavailable (this is
backwards incompatible).

Similarly for all channels, the `channels.*.pythonPackages` attribute has to
be used. The version that `pythonPackages` and co. is set to is currently
just hardcoded in the flox entry-point, except that it is set to the correct
version for package definitions in pythonPackages/*.

In future commits, this will be more configurable, but for now this commit
lays the foundation to support better version propagation.