Allow package set version configuration with config.nix files
This introduces the config.nix files for package definitions. It is used
to declare that the package definition needs access to certain package sets.

This means that for e.g. a package definition in pythonPackages/foo, in
order for the default.nix file to receive the `pythonPackages` argument,
it needs to specify a config.nix file with contents

  {
    packageSets.python.type = "lib";
  }

Currently this is needed, in the future pythonPackages/* will have such
a statement by default without the user having to write one.

A type of "lib" means that the version is inherited from the importing
channel. The alternative is the "app" type, which allows the package to
choose its own version:

  {
    packageSets.python.type = "app";
    packageSets.python.app.version = "3.9.0";
  }

This should only be used when the application doesn't propagate any
python modules, such as with buildPythonApplication, and ensures that
the whole tree of dependencies uses the specified version.