Introduce _floxPathDepth for accessing _floxPath with eval failures
When the flox tool is evaluating an attribute path for its package file via e.g.

    nix-instantiate -A pythonPackages.foo._floxPath

it can happen that there is a syntax error in the expression, leading to
even `_floxPath` not being successfully evaluated.

Note that only the features of `meta.importNix` (allowing delegation of a
Nix definition to a source repository), which sets `_floxPath` to the source
repository file itself, and the feature of packages being able to set a
different `_floxPath` via an output attribute directly, are the reason
this is problematic at all. Because without these features, the mapping
from attribute path to file is trivially always directly into the path
matching the attribute in the channel itself.

In order to still be able to get a `_floxPath` out of the channel in
presence of these features, a new internal `_floxPathDepth` argument is
introduced. With its current default value of 2, it behaves the same as
normal, evaluations errors on any level cause it to fail.

When set to 1 however, it will essentially ignore the result of any
`_floxPath`'s returned from files imported via `meta.importNix`. This
means that if there's a syntax error in a source repository imported via
`meta.importNix`, `meta.importNix`'s own `_floxPath` can still evaluate
to the file giving the syntax error.

    nix-instantiate -A pythonPackages.foo._floxPath --arg _floxPathDepth 1

Similarly when set to 0, it will already ignore any `_floxPath`'s
returned from the channel's packages directly. This means that if
there's a syntax error in `pythonPackages/foo/default.nix`, we will
still get a `_floxPath` that points to that file.

    nix-instantiate -A pythonPackages.foo._floxPath --arg _floxPathDepth 0