Refactor for a single channel layer
* Motivation

This is essentially a complete rewrite of how flox channels are handled. The
problem that's being solved here is that previously it was very easy to get
package conflict problems between channels at runtime.

Previously, all channels were treated completely separately: They had their
own nixpkgs version and overlays, and their own packages. A channel
could completely determine on its own where dependencies come from, even
when it is used by another channel.

This can cause problems, because what if channel A specifies a version
of package X, X_A, and uses this for all its packages. However, channel
A also imports and uses packages from channel B, which can define its
own version of X, X_B. Now in A's resulting packages, you can easily end
up with both X_A and X_B, which can give runtime conflicts and
inconsistencies.

The previous deep-override mechanism implementation was thought to be a good
solution to this, since it allowed each channel to essentially define a
nixpkgs overlay, where the root channels overlay is applied last, and
each channel that is being depended on prepends its own overlay to the
ones from the channel it is being imported from.

However that didn't end up working very well, because each channel could
end up with a different set of overlays, therefore potentially different
versions of all nixpkgs packages. Also, these overlays are only applied
to nixpkgs itself, not the flox channels. This means you can end up with
the same package name in multiple channels but with a different value.

* Solution

To fix this, instead of treating all channels as separate, we merge them
all together into a single package set. Doing this allows us to
explicitly deal with conflicts ahead-of-time.

Notably, we rely on each channel having declared its dependency channels
properly, at which point we can find the transitive closure of necessary
channels. We then merge all package declarations from all
channels into a single set. If there are any conflicts, the root channel
(the channel we are evaluating itself) will have to specify how to
resolve them. Packages defined in the root channel notably always take
precedence over non-root channels.

Now that we know where each package comes from, we can use this to
create a scope of called packages. And we notably call each package,
_even ones not in the root channel_, using this scope determined by the
root channel. This means that when a package from an arbitrary channel
is called, depending on the conflict resolution provided in the root
channel, it might be called with arguments entirely different than what
they would be if the arbitrary channel itself were the root channel.

And this is what makes it work: It ensures that for every package, in every
channel, which uses a certain argument in its expression file, the
argument value is the same for all of them.

Note that this can have false positives and negatives though, because if
multiple versions of the same package exist with different attribute
names, it of course won't give a merge conflict, even if both are used.
Similarly, there can be conflicts that occur just because two channels
define a package under the same name, even if they're completely
unrelated. This is a reasonable tradeoff however, and this notably also
mirrors how nixpkgs itself works.