channel/tests: Add more tests and fix existing ones
With the previous commit, some tests started failing with the
deprecation of some features, specifically:
- `extraOverlays` isn't (and can't be) supported anymore
- The channel discovery tests don't really make sense anymore since we
  aren't listing all `NIX_PATH` entries anymore, but require people to
  explicitly declare dependencies with `channels.json`

These were removed.

In addition, a number of new tests were added for:
- Ensuring that conflict resolution works, both at the root and
  override/super level
- Ensuring that dependency specification works
- Ensuring that all the newly added utilities work

In addition, this commit contains a small change in the testing
framework that allows it to specify a shared pregenerated
package-sets.json, speeding up tests considerably.