Make builders propagate to the nixpkgs equivalent if no project arg passed
With the previous refactoring for a single channel layer, the flox
channel implicitly overrides the nixpkgs channel, since they both define
a number of the same builders. However this also means that the nixpkgs
builder isn't accessible directly anymore. So what we do to fix this is
to just make the flox builder be the nixpkgs builder if no flox-specific
project argument is provided. This makes the flox builder essentially a
superset in functionality.