Changes for flox(pkgs) -> flox-lib renaming
From now on, the flox channel is now the flox-lib channel. It contains
all the Nix code to make channels work together, and all the standard
builders. It will newly be added as a special entry to NIX_PATH.

This allows using the `flox` channel namespace for actual flox packages.